pyproject.toml
[tool.poetry]
name = "firstdiscordbot"
version = "0.1.0"
description = "My first Discord bot"
authors = ["Parker Wahle <regulad@regulad.xyz>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
